+++ src/appsettings.h
@@ -3,7 +3,7 @@
 
 #include <QSettings>
 #include "cor/light.h"
-
+#include <QObject>
 
 /*!
  * \copyright
@@ -14,13 +14,15 @@
  * \brief The ProtocolSettings class manages which CommTypes are enabled and disabled. It manages the settings
  *        between application sessions by saving to a QSettings instance.
  */
-class ProtocolSettings
+class AppSettings : QObject
 {
+    Q_OBJECT
+
 public:
     /*!
      * \brief ProtocolSettings constructor
      */
-    ProtocolSettings();
+    AppSettings();
 
     //------------------
     // Enable/Disable CommTypes
@@ -44,6 +46,36 @@ public:
     bool enabled(EProtocolType type);
 
     //------------------
+    // Timeouts
+    //------------------
+
+    /*!
+     * \brief timeOut getter for the amount of minutes it takes for the LEDs
+     *        to "time out." When this happens, they turn off, saving you
+     *        electricity!
+     * \return the time it'll take for the LEDs to time out.
+     */
+    int timeout() { return mTimeout;}
+
+    /*!
+     * \brief enableTimeout true to enable timeout modes, false to disable them. If they are enabled,
+     *        all lights will turn off after their timeoutInterval has passed if no new packets are sent
+     *        to them
+     * \param timeout true to enable timeout modes, false to disable them.
+     */
+    void enableTimeout(bool timeout);
+
+    /// true if timeouts are enabled globally, false otherwise.
+    bool timeoutEnabled() { return mTimeoutEnabled; }
+
+    /*!
+     * \brief updateTimeout update how many minutes it takes for lights to turn themselves off automatically.
+     *        Use a value of 0 to keep lights on indefinitely (until you unplug them or change the setting).
+     * \param timeout the new number of minutes it takes for LEDs to time out and turn off.
+     */
+    void updateTimeout(int timeout);
+
+    //------------------
     // Miscellaneous
     //------------------
 
@@ -52,6 +84,15 @@ public:
 
     /// getter for count of ProtocolSettings enabled
     uint32_t numberEnabled();
+
+signals:
+
+    /*!
+     * \brief settingsUpdate there has been an update to the settings such as when to timeout or the speed
+     *        of routines.
+     */
+    void settingsUpdate();
+
 private:
     /*!
      * \brief mProtocolsInUse a vector of bools that store which CommTypes are currently
@@ -63,6 +104,12 @@ private:
      * \brief mSettings pointer to QSettings, used to store and access data in persistent app memory.
      */
     QSettings *mSettings;
+
+    /// true if lights should turn off after X hours of no use, false othwerise.
+    bool mTimeoutEnabled;
+
+    /// value for how long lights should stay on before timeout used globally across all lights
+    int mTimeout;
 };
 
 #endif // ProtocolSettings_H