+++ src/devicelist.h
@@ -10,7 +10,7 @@
 #include "cor/protocols.h"
 #include "cor/palette.h"
 #include "comm/commhue.h"
-#include "comm/protocolsettings.h"
+#include "appsettings.h"
 
 /*!
  * \copyright
@@ -55,14 +55,6 @@ public:
     int brightness();
 
     /*!
-     * \brief timeOut getter for the amount of minutes it takes for the LEDs
-     *        to "time out." When this happens, they turn off, saving you
-     *        electricity!
-     * \return the time it'll take for the LEDs to time out.
-     */
-    int timeout() { return mTimeout;}
-
-    /*!
      * \brief palette getter for the current palette.
      * \return the EPalette that represents the colors being displayed on
      *         the LED array.
@@ -88,24 +80,6 @@ public:
      * \param on true if you want to turn on, false if you want to turn off.
      */
     void turnOn(bool on);
-
-    /*!
-     * \brief enableTimeout true to enable timeout modes, false to disable them. If they are enabled,
-     *        all lights will turn off after their timeoutInterval has passed if no new packets are sent
-     *        to them
-     * \param timeout true to enable timeout modes, false to disable them.
-     */
-    void enableTimeout(bool timeout);
-
-    /// true if timeouts are enabled globally, false otherwise.
-    bool timeoutEnabled() { return mTimeoutEnabled; }
-
-    /*!
-     * \brief updateTimeout update how many minutes it takes for lights to turn themselves off automatically.
-     *        Use a value of 0 to keep lights on indefinitely (until you unplug them or change the setting).
-     * \param timeout the new number of minutes it takes for LEDs to time out and turn off.
-     */
-    void updateTimeout(int timeout);
 
     /*!
      * \brief isOn true if any device is on, false if all are off.
@@ -255,12 +229,6 @@ signals:
      */
     void dataUpdate();
 
-    /*!
-     * \brief settingsUpdate there has been an update to the settings such as when to timeout or the speed
-     *        of routines.
-     */
-    void settingsUpdate();
-
 private:
 
     /*!
@@ -270,11 +238,6 @@ private:
      */
     std::list<cor::Light> mDevices;
 
-    /// true if lights should turn off after X hours of no use, false othwerise.
-    bool mTimeoutEnabled;
-
-    /// value for how long lights should stay on before timeout used globally across all lights
-    int mTimeout;
 };
 
 #endif // DATALAYER_H